Versions_update  Statistik
Statistik zur ui hinzugefügt (gesehen/nicht gesehen im % verhältniss zur anzahl aller fakten

Versions Udate:

Passwörter und sensiebele Daten in config ausgelagert und "verschleiert" um Projekt Github tauglich zu machen.
