--- src/fakten.csv
@@ -1,11 +1,0 @@
-facts
-Daylight Saving Time is not observed in most of the state of Arizona and parts of Indiana.
-The average American butt is 14.9 inches long.
-Over 1000 birds a year die from smashing into windows!
-"The crack of a whip is actually a tiny sonic boom, since the tip breaks the sound barrier."
-"In 1986 Congress & President Ronald Reagan signed Public Law 99-359, which changed Daylight Saving Time from the last Sunday in April to the first Sunday in April.  It was estimated to save the nation about 300,000 barrels of oil each year by adding most of the month April to D.S.T."
-The muzzle of a lion is like a fingerprint - no two lions have the same pattern of whiskers.  
-The dot over the letter `i` is called a tittle.  
-"Thomas Edison, acclaimed inventor of the light bulb, was afraid of the dark."
-A standard slinky measures 87 feet when stretched out.
-Worcestershire sauce in essentially an Anchovy Ketchup.
