+++ src/gruppenarbeit/Präsentation_Version/2.API_csv_export_by_birol.py
@@ -3,16 +3,14 @@ import csv
 import os
 import time
 
-from config import CSV_PATH
 
-# API_URL
 url = 'https://uselessfacts.jsph.pl/api/v2/facts/random'
 
-# Anzahl der Fakten die gesucht werden sollen
-num_facts = 10
+num_facts = 50
+folder_path = r'/src'
+file_path = os.path.join(folder_path, 'fakten.csv')
 
-file_path = CSV_PATH
-os.makedirs(os.path.dirname(CSV_PATH), exist_ok=True)
+os.makedirs(folder_path, exist_ok=True)
 facts_set = set()
 
 # CSV erstellen ohne Anführungszeichen
